feat: add ContentTypeSidebar, EntryTable, SortMenu, TableHeader, and TableRow components with pagination and sorting functionality
