+++ apps/bulk-edit/src/locations/Page/components/EntryTable.tsx
@@ -1,0 +1,70 @@
+import React from 'react';
+import { Table, Box, Pagination } from '@contentful/f36-components';
+import { Entry, ContentTypeField } from '../types';
+import { ContentTypeProps } from 'contentful-management';
+import { styles } from '../styles';
+import { TableHeader } from './TableHeader';
+import { TableRow } from './TableRow';
+
+interface EntryTableProps {
+  entries: Entry[];
+  fields: ContentTypeField[];
+  contentType?: ContentTypeProps;
+  spaceId: string;
+  environmentId: string;
+  locale: string;
+  activePage: number;
+  totalEntries: number;
+  itemsPerPage: number;
+  onPageChange: (page: number) => void;
+  onItemsPerPageChange: (itemsPerPage: number) => void;
+  pageSizeOptions: number[];
+}
+
+export const EntryTable: React.FC<EntryTableProps> = ({
+  entries,
+  fields,
+  contentType,
+  spaceId,
+  environmentId,
+  locale,
+  activePage,
+  totalEntries,
+  itemsPerPage,
+  onPageChange,
+  onItemsPerPageChange,
+  pageSizeOptions,
+}) => {
+  return (
+    <>
+      <Table testId="bulk-edit-table" style={styles.table}>
+        <TableHeader fields={fields} />
+        <Table.Body>
+          {entries.map((entry) => (
+            <TableRow
+              key={entry.sys.id}
+              entry={entry}
+              fields={fields}
+              contentType={contentType}
+              spaceId={spaceId}
+              environmentId={environmentId}
+              locale={locale}
+            />
+          ))}
+        </Table.Body>
+      </Table>
+      <Box marginTop="spacingM">
+        <Pagination
+          activePage={activePage}
+          onPageChange={onPageChange}
+          totalItems={totalEntries}
+          showViewPerPage
+          viewPerPageOptions={pageSizeOptions}
+          itemsPerPage={itemsPerPage}
+          onViewPerPageChange={onItemsPerPageChange}
+          aria-label="Pagination navigation"
+        />
+      </Box>
+    </>
+  );
+};
